+++ plugins/main/public/components/overview/compliance-table/components/subrequirements/subrequirements.tsx
@@ -89,6 +89,7 @@ export class ComplianceSubrequirements extends Component {
       nist: 'rule.compliance.nist_800_53',
       hipaa: 'rule.compliance.hipaa',
       tsc: 'rule.compliance.tsc',
+      cmmc: 'rule.compliance.cmmc',
       fedramp: 'rule.compliance.fedramp',
       nis2: 'rule.compliance.nis2',
     };