+++ plugins/main/public/components/overview/index.ts
@@ -13,6 +13,7 @@ export { DashboardNIS2 } from './regulatory-compliance/nis2/dashboards';
 export { DashboardOffice365 } from './office/dashboard';
 export { DashboardThreatHunting } from './threat-hunting/dashboard';
 export { DashboardTSC } from './regulatory-compliance/tsc/dashboards';
-export { DashboardFEDRAMP } from './regulatory-compliance/fedramp/dashboards';
+export { DashboardCMMC } from './regulatory-compliance/cmmc/dashboards';
+export { DashboardFedRAMP } from './regulatory-compliance/fedramp/dashboards';
 export { DashboardVuls, InventoryVuls } from './vulnerabilities';
 export { DashboardAzure } from './azure/dashboards';